fix: step 2.1
