+++ test/e2e/nightfall-sdk.test.mjs
@@ -285,90 +285,6 @@ describe('Testing the Nightfall SDK', () => {
     await checkUserBalances();
   });
 
-
-  describe('Basic Proposer tests', () => {
-    it('should register a proposer', async () => {
-      let proposers;
-      ({ proposers } = await nf3Proposer2.getProposers());
-      // we have to pay 10 ETH to be registered
-      const startBalance = await getBalance(nf3Proposer2.ethereumAddress);
-      const res = await nf3Proposer2.registerProposer();
-      stateBalance += bond;
-      expectTransaction(res);
-      ({ proposers } = await nf3Proposer2.getProposers());
-      const endBalance = await getBalance(nf3Proposer2.ethereumAddress);
-      expect(endBalance - startBalance).to.closeTo(-bond, gasCosts);
-      const thisProposer = proposers.filter(p => p.thisAddress === nf3Proposer2.ethereumAddress);
-      expect(thisProposer.length).to.be.equal(1);
-    });
-
-    it('should register other proposer', async () => {
-      let proposers;
-      ({ proposers } = await nf3Proposer3.getProposers());
-      // we have to pay 10 ETH to be registered
-      const startBalance = await getBalance(nf3Proposer3.ethereumAddress);
-      const res = await nf3Proposer3.registerProposer();
-      stateBalance += bond;
-      expectTransaction(res);
-      ({ proposers } = await nf3Proposer3.getProposers());
-      const endBalance = await getBalance(nf3Proposer3.ethereumAddress);
-      expect(endBalance - startBalance).to.closeTo(-bond, gasCosts);
-      const thisProposer = proposers.filter(p => p.thisAddress === nf3Proposer3.ethereumAddress);
-      expect(thisProposer.length).to.be.equal(1);
-    });
-
-    it('should de-register a proposer', async () => {
-      let proposers;
-      ({ proposers } = await nf3Proposer1.getProposers());
-      let thisProposer = proposers.filter(p => p.thisAddress === nf3Proposer1.ethereumAddress);
-      expect(thisProposer.length).to.be.equal(1);
-      const res = await nf3Proposer1.deregisterProposer();
-      expectTransaction(res);
-      ({ proposers } = await nf3Proposer1.getProposers());
-      thisProposer = proposers.filter(p => p.thisAddress === nf3Proposer1.ethereumAddress);
-      expect(thisProposer.length).to.be.equal(0);
-    });
-
-    it('Should create a failing withdrawBond (because insufficient time has passed)', async () => {
-      let error = null;
-      try {
-        await nf3Proposer1.withdrawBond();
-      } catch (err) {
-        error = err;
-      }
-      expect(error.message).to.satisfy(
-        message =>
-          message.includes(
-            'Returned error: VM Exception while processing transaction: revert It is too soon to withdraw your bond',
-          ) || message.includes('Transaction has been reverted by the EVM'),
-      );
-    });
-
-    it('Should create a passing withdrawBond (because sufficient time has passed)', async () => {
-      if (nodeInfo.includes('TestRPC')) await timeJump(3600 * 24 * 10); // jump in time by 7 days
-      if (nodeInfo.includes('TestRPC')) {
-        const res = await nf3Proposer1.withdrawBond();
-        expectTransaction(res);
-      } else {
-        let error = null;
-        try {
-          await nf3Proposer1.withdrawBond();
-        } catch (err) {
-          error = err;
-        }
-        expect(error.message).to.include('Transaction has been reverted by the EVM');
-      }
-    });
-
-    after(async () => {
-      // After the proposer tests, re-register proposers
-      await nf3Proposer2.deregisterProposer();
-      await nf3Proposer3.deregisterProposer();
-      await nf3Proposer1.registerProposer();
-      stateBalance += bond;
-    });
-  });
-
   describe('Basic Challenger tests', () => {
     it('should register a challenger', async () => {
       const res = await nf3Challenger.registerChallenger();
