+++ test/e2e/signingKeys.json
@@ -1,0 +1,10 @@
+{
+  "walletTest": "0xd42905d0582c476c4b74757be6576ec323d715a0c7dcff231b6348b7ab0190eb",
+  "user1": "0x4775af73d6dc84a0ae76f8726bda4b9ecf187c377229cb39e1afa7a18236a69e",
+  "user2": "0xd42905d0582c476c4b74757be6576ec323d715a0c7dcff231b6348b7ab0190eb",
+  "proposer1": "0x4775af73d6dc84a0ae76f8726bda4b9ecf187c377229cb39e1afa7a18236a69d",
+  "proposer2": "0xd42905d0582c476c4b74757be6576ec323d715a0c7dcff231b6348b7ab0190eb",
+  "proposer3": "0xfbc1ee1c7332e2e5a76a99956f50b3ba2639aff73d56477e877ef8390c41e0c6",
+  "challenger": "0xd42905d0582c476c4b74757be6576ec323d715a0c7dcff231b6348b7ab0190eb",
+  "liquidityProvider": "0xfbc1ee1c7332e2e5a76a99956f50b3ba2639aff73d56477e877ef8390c41e0c6"
+}
